+++ clang-tools-extra/clang-tidy/modernize/UseUsingCheck.cpp
@@ -25,107 +25,88 @@ void UseUsingCheck::registerMatchers(MatchFinder *Finder) {
     return;
   Finder->addMatcher(typedefDecl(unless(isInstantiated())).bind("typedef"),
                      this);
-}
-
-// Checks if 'typedef' keyword can be removed - we do it only if
-// it is the only declaration in a declaration chain.
-static bool CheckRemoval(SourceManager &SM, SourceLocation StartLoc,
-                         ASTContext &Context) {
-  assert(StartLoc.isFileID() && "StartLoc must not be in a macro");
-  std::pair<FileID, unsigned> LocInfo = SM.getDecomposedLoc(StartLoc);
-  StringRef File = SM.getBufferData(LocInfo.first);
-  const char *TokenBegin = File.data() + LocInfo.second;
-  Lexer DeclLexer(SM.getLocForStartOfFile(LocInfo.first), Context.getLangOpts(),
-                  File.begin(), TokenBegin, File.end());
-
-  Token Tok;
-  int NestingLevel = 0; // Parens, braces, and square brackets
-  int AngleBracketLevel = 0;
-  bool FoundTypedef = false;
-
-  while (!DeclLexer.LexFromRawLexer(Tok) && !Tok.is(tok::semi)) {
-    switch (Tok.getKind()) {
-    case tok::l_brace:
-      if (NestingLevel == 0 && AngleBracketLevel == 0) {
-        // At top level, this might be the `typedef struct {...} T;` case.
-        // Inside parens, square brackets, or angle brackets it's not.
-        return false;
-      }
-      ++NestingLevel;
-      break;
-    case tok::l_paren:
-    case tok::l_square:
-      ++NestingLevel;
-      break;
-    case tok::r_brace:
-    case tok::r_paren:
-    case tok::r_square:
-      --NestingLevel;
-      break;
-    case tok::less:
-      // If not nested in paren/brace/square bracket, treat as opening angle bracket.
-      if (NestingLevel == 0)
-        ++AngleBracketLevel;
-      break;
-    case tok::greater:
-      // Per C++ 17 Draft N4659, Section 17.2/3
-      //   https://timsong-cpp.github.io/cppwp/n4659/temp.names#3:
-      // "When parsing a template-argument-list, the first non-nested > is
-      // taken as the ending delimiter rather than a greater-than operator."
-      // If not nested in paren/brace/square bracket, treat as closing angle bracket.
-      if (NestingLevel == 0)
-        --AngleBracketLevel;
-      break;
-    case tok::comma:
-      if (NestingLevel == 0 && AngleBracketLevel == 0) {
-        // If there is a non-nested comma we have two or more declarations in this chain.
-        return false;
-      }
-      break;
-    case tok::raw_identifier:
-      if (Tok.getRawIdentifier() == "typedef") {
-        FoundTypedef = true;
-      }
-      break;
-    default:
-      break;
-    }
-  }
-
-  // Sanity check against weird macro cases.
-  return FoundTypedef;
+  // This matcher used to find structs defined in source code within typedefs.
+  // They appear in the AST just *prior* to the typedefs.
+  Finder->addMatcher(cxxRecordDecl(unless(isImplicit())).bind("struct"), this);
 }
 
 void UseUsingCheck::check(const MatchFinder::MatchResult &Result) {
+  // Match CXXRecordDecl only to store the range of the last non-implicit full
+  // declaration, to later check whether it's within the typdef itself.
+  const auto *MatchedCxxRecordDecl =
+      Result.Nodes.getNodeAs<CXXRecordDecl>("struct");
+  if (MatchedCxxRecordDecl) {
+    LastCxxDeclRange = MatchedCxxRecordDecl->getSourceRange();
+    return;
+  }
+
   const auto *MatchedDecl = Result.Nodes.getNodeAs<TypedefDecl>("typedef");
   if (MatchedDecl->getLocation().isInvalid())
     return;
-
-  auto &Context = *Result.Context;
-  auto &SM = *Result.SourceManager;
 
   SourceLocation StartLoc = MatchedDecl->getBeginLoc();
 
   if (StartLoc.isMacroID() && IgnoreMacros)
     return;
 
-  auto Diag = diag(StartLoc, "use 'using' instead of 'typedef'");
+  static const char *UseUsingWarning = "use 'using' instead of 'typedef'";
 
-  // do not fix if there is macro or array
-  if (MatchedDecl->getUnderlyingType()->isArrayType() || StartLoc.isMacroID())
+  // Warn at StartLoc but do not fix if there is macro or array.
+  if (MatchedDecl->getUnderlyingType()->isArrayType() || StartLoc.isMacroID()) {
+    diag(StartLoc, UseUsingWarning);
     return;
-
-  if (CheckRemoval(SM, StartLoc, Context)) {
-    auto printPolicy = PrintingPolicy(getLangOpts());
-    printPolicy.SuppressScope = true;
-    printPolicy.ConstantArraySizeAsWritten = true;
-    printPolicy.UseVoidForZeroParams = false;
-
-    Diag << FixItHint::CreateReplacement(
-        MatchedDecl->getSourceRange(),
-        "using " + MatchedDecl->getNameAsString() + " = " +
-            MatchedDecl->getUnderlyingType().getAsString(printPolicy));
   }
+
+  auto printPolicy = PrintingPolicy(getLangOpts());
+  printPolicy.SuppressScope = true;
+  printPolicy.ConstantArraySizeAsWritten = true;
+  printPolicy.UseVoidForZeroParams = false;
+
+  std::string Type = MatchedDecl->getUnderlyingType().getAsString(printPolicy);
+  std::string Name = MatchedDecl->getNameAsString();
+  SourceRange ReplaceRange = MatchedDecl->getSourceRange();
+
+  // typedefs with multiple comma-separated definitions produce multiple
+  // consecutive TypedefDecl nodes whose SourceRanges overlap. Each range starts
+  // at the "typedef" and then continues *across* previous definitions through
+  // the end of the current TypedefDecl definition.
+  std::string Using = "using ";
+  if (ReplaceRange.getBegin().isMacroID() ||
+      ReplaceRange.getBegin() >= LastReplacementEnd) {
+    // This is the first (and possibly the only) TypedefDecl in a typedef. Save
+    // Type and Name in case we find subsequent TypedefDecl's in this typedef.
+    FirstTypedefType = Type;
+    FirstTypedefName = Name;
+  } else {
+    // This is additional TypedefDecl in a comma-separated typedef declaration.
+    // Start replacement *after* prior replacement and separate with semicolon.
+    ReplaceRange.setBegin(LastReplacementEnd);
+    Using = ";\nusing ";
+
+    // If this additional TypedefDecl's Type starts with the first TypedefDecl's
+    // type, make this using statement refer back to the first type, e.g. make
+    // "typedef int Foo, *Foo_p;" -> "using Foo = int;\nusing Foo_p = Foo*;"
+    if (Type.size() > FirstTypedefType.size() &&
+        Type.substr(0, FirstTypedefType.size()) == FirstTypedefType)
+      Type = FirstTypedefName + Type.substr(FirstTypedefType.size() + 1);
+  }
+  if (!ReplaceRange.getEnd().isMacroID())
+    LastReplacementEnd = ReplaceRange.getEnd().getLocWithOffset(Name.size());
+
+  auto Diag = diag(ReplaceRange.getBegin(), UseUsingWarning);
+
+  // If typedef contains a full struct/class declaration, extract its full text.
+  if (LastCxxDeclRange.isValid() && ReplaceRange.fullyContains(LastCxxDeclRange)) {
+    bool Invalid;
+    Type =
+        Lexer::getSourceText(CharSourceRange::getTokenRange(LastCxxDeclRange),
+                             *Result.SourceManager, getLangOpts(), &Invalid);
+    if (Invalid)
+      return;
+  }
+
+  std::string Replacement = Using + Name + " = " + Type;
+  Diag << FixItHint::CreateReplacement(ReplaceRange, Replacement);
 }
 
 } // namespace modernize